% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerationBands.R
\name{accelerationBands}
\alias{accelerationBands}
\alias{ABands}
\title{Acceleration Bands}
\usage{
accelerationBands(HLC, n = 20, maType, ...)
}
\arguments{
\item{HLC}{Object that is coercible to xts or matrix and contains High-Low-Close prices.}

\item{n}{Number of periods for moving average.}

\item{maType}{A function or a string naming the function to be called.}

\item{\dots}{Other arguments to be passed to the \code{maType} function.}
}
\value{
A object of the same class as \code{HLC} or a matrix (if
\code{try.xts} fails) containing the columns:
 \describe{
  \item{ dn }{ The lower Acceleration Band. }
  \item{ mavg }{ The middle Moving Average (see notes). }
  \item{ up }{ The upper Acceleration Band. }
 }
}
\description{
Acceleration Bands
}
\examples{

 ## The examples below show the differences between using a
 ## High-Low-Close series when
 ## calculating Acceleration Bands.
 data(ttrc)
 accelerationBands.HLC <- accelerationBands( ttrc[,c("High","Low","Close")] )

}
\references{
The following site(s) were used to code/document this indicator:

\cr \url{https://quantstrategy.io/blog/understanding-the-acceleration-bands/} \cr
\cr \url{https://trendspider.com/learning-center/getting-started-with-acceleration-bands-in-technical-analysis/} \cr
\cr \url{https://github.com/TA-Lib/ta-lib/blob/main/src/ta_func/ta_ACCBANDS.c} \cr
}
\seealso{
See \code{\link{EMA}}, \code{\link{SMA}}, etc. for moving average
options; and note Warning section.
}
\author{
Serkan Korkmaz
}
\keyword{ts}
