% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttthemes.R
\name{get_font}
\alias{get_font}
\title{Function to get default fonts for \code{ttr::theme_tt()}}
\usage{
get_font(type = "slide")
}
\arguments{
\item{type}{One of \code{"slide"}, \code{"paper"}, \code{"slide_title"}, \code{"paper_title"}}
}
\value{
A font name (e.g., \code{"source-sans-pro"}, \code{"alegreya-sans"}, \code{"lmroman"}, \code{"tgheros"})
}
\description{
Return font name of font \code{type}
}
