% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttthemes.R
\docType{data}
\name{paper_font}
\alias{paper_font}
\title{Default font for papers}
\format{
An object of class \code{character} of length 1.
}
\usage{
paper_font
}
\description{
This font is automatically imported with \code{ttr}: Latin Modern Roman
}
\keyword{datasets}
