% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttthemes.R
\name{set_font}
\alias{set_font}
\title{Function to set default fonts for \code{ttr::theme_tt()}}
\usage{
set_font(type = "slide", font)
}
\arguments{
\item{type}{One of \code{"slide"}, \code{"paper"}, \code{"slide_title"}, \code{"paper_title"}}

\item{font}{A font name (e.g., \code{"source-sans-pro"}, \code{"alegreya-sans"}, \code{"lmroman"}, \code{"tgheros"})}
}
\description{
Change font \code{type} (\code{"slide"}, \code{"paper"}, \code{"slide_title"}, \code{"paper_title"}) to \code{font}
}
