% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttthemes.R
\docType{data}
\name{slide_font}
\alias{slide_font}
\title{Default font for slides}
\format{
An object of class \code{character} of length 1.
}
\usage{
slide_font
}
\description{
This font is automatically imported with \code{ttr}: TeX Gyre Heros
}
\keyword{datasets}
