% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs_plots.R
\name{specs_coef_plot}
\alias{specs_coef_plot}
\title{Draw a coefficient plot.}
\usage{
specs_coef_plot(
  coef_grid,
  palette = "green",
  style = "slide",
  coef_ylabel = "Coefficient estimate",
  coef_ylim = NULL,
  point_size = NULL,
  error_alpha = NULL,
  error_geom = NULL,
  error_width = 0,
  hline = NULL,
  pcolors = NULL,
  standalone = FALSE
)
}
\arguments{
\item{coef_grid}{A dataframe with the following components:
\code{index} (if part of a specification plot) or \code{coef_name} (if standalone),
\code{coef}, \code{error_low}, \code{error_high}, \code{p} for p-value categories
(character with values, default should be \code{"p<0.01"}, \code{"p<0.05"}, \code{"p<0.10"}, \verb{p>0.10"})
or \code{pval} for the raw numeric p-values.}

\item{palette}{Change the palette.
Default to \code{"green"} for green-shaded plot.
Change to \code{"brown"} for brown-shaded plot.}

\item{style}{Change the theme style.
Default to \code{"slide"} for default slide fonts and themed axis title and text.
Change to \code{"paper"} for default paper fonts and black axis title and text.}

\item{coef_ylabel}{A string specifying the y-axis label on the coefficient panel.
Default to `"Coefficient estimate"'.}

\item{coef_ylim}{A numeric vector of length two indicating the
minimum and maximum values of the y-axis in the coefficient plot.
Default to \code{NULL} to use \code{ggplot2} default.}

\item{point_size}{A numeric scalar indicating the size of the coefficient estimate points.
Default to \code{NULL} and set according to the number of specs.}

\item{error_alpha}{A numeric scalar indicating the alpha of the error geom.
Default to \code{NULL} and set to \code{0.5} for \code{"errorbar"} and \code{0.21} for \code{"ribbon"}.}

\item{error_geom}{A string indicating the type of geom that should be used to
depict confidence intervals on coefficient estimates.
Currently supported are \code{"ribbon"}, \code{"errorbar"}, and \code{"none"}.
Default to \code{NULL} and set to "errorbar"\verb{if fewer than 100 specs and to}"ribbon"` otherwise.}

\item{error_width}{A numeric scalar indicating the width of the error bar.
Default to 0 and only applicable when \code{error_geom} is set to \code{"errorbar"}.}

\item{hline}{A numeric scalar indicating a horizontal line at value \code{hline}.
Default to \code{NULL} for none.}

\item{pcolors}{A named vector of colors for points by \code{p} value
(defaults are "p<0.01", "p<0.05", "p<0.10", "p>0.10").
Default to \code{NULL} for default colors based on \code{ttr::ttcolors}.}

\item{standalone}{A boolean to indicate whether this is a standalone plot or
part of a specification plot with a control grid panel.
Default to `FALSE'.}
}
\value{
A \code{ggplot2} object depicting the coefficient plot
}
\description{
\code{specs_coef_plot} produces the coefficient plot in a specification plot or standalone.
}
