% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs_plots.R
\name{specs_combine_plots}
\alias{specs_combine_plots}
\title{Combine the coefficient and control grid plots.}
\usage{
specs_combine_plots(
  coef_plot,
  control_plot,
  ncoefs,
  ratio = 2,
  control_spacing = NULL
)
}
\arguments{
\item{coef_plot}{A \code{ggplot2} plot: coefficient plot for the upper panel.}

\item{control_plot}{A \code{ggplot2} plot: control grid plot for the lower panel.}

\item{ncoefs}{A numeric scalar indicating how many specifications there are.}

\item{ratio}{A numeric scalar indicating the height ratio of the
coefficient plot relative to the control grid plot.}

\item{control_spacing}{A numeric scalar indicating how large the control geoms
should be.
For \code{control_geom=="circle"}, this is the diameter of the circle.
For \code{control_geom=="rect"}, this is the width of the rectangle.
Default to \code{NULL} and set to \code{0.75} if fewer than 40 specs and 1 otherwise.}
}
\value{
A \code{ggplot2} object depicting the specification plot
}
\description{
\code{specs_combine_plots} produces the combined specification plot.
}
