% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs_plots.R
\name{specs_control_plot}
\alias{specs_control_plot}
\title{Draw a control grid plot.}
\usage{
specs_control_plot(
  control_grid,
  palette = "green",
  style = "slide",
  control_geom = "rect",
  control_spacing = NULL,
  control_text_size = 9,
  trim_top = 0
)
}
\arguments{
\item{control_grid}{A dataframe with \code{index} (of the specs), \code{key} (name of a spec key) and \code{value} (0 or 1).}

\item{palette}{Change the palette.
Default to \code{"green"} for green-shaded plot.
Change to \code{"brown"} for brown-shaded plot.}

\item{style}{Change the theme style.
Default to \code{"slide"} for default slide fonts and themed axis title and text.
Change to \code{"paper"} for default paper fonts and black axis title and text.}

\item{control_geom}{A string indicating the geom that should be used to
indicate the presence of controls.
Currently supported are \code{"circle"} and \code{"rect"}.
Default to \code{"rect"}.}

\item{control_spacing}{A numeric scalar indicating how large the control geoms
should be.
For \code{control_geom=="circle"}, this is the diameter of the circle.
For \code{control_geom=="rect"}, this is the width of the rectangle.
Default to \code{NULL} and set to \code{0.75} if fewer than 40 specs and \code{1} otherwise.}

\item{control_text_size}{A numeric scalar indicating how large the
control name text should be.
Default to 9.}

\item{trim_top}{A numeric scalar indicating how close the bottom panel
(displaying presence of controls) should be to the
top panel (displaying presence of coefficients).
Useful when dealing with large CIs.
Default to \code{0}.}
}
\value{
A \code{ggplot2} object depicting the control grid plot
}
\description{
\code{specs_control_plot} produces the control grid plot in a specification plot.
}
