% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttcolors.R
\name{ttcolor}
\alias{ttcolor}
\title{Function to extract tt colors as hex codes}
\usage{
ttcolor(...)
}
\arguments{
\item{...}{Character names of ttcolors}
}
\description{
Each color name will be matched with a corresponding hex code.
There can be repeated names.
}
