% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttcolors.R
\name{ttpalette}
\alias{ttpalette}
\title{ttpalette(n) function}
\usage{
ttpalette(palette = "main", reverse = FALSE, nmax = NULL, order = NULL, ...)
}
\arguments{
\item{palette}{Character name of palette. Run \code{names(ttpalettes)} for available options.}

\item{reverse}{If TRUE, reverses the order of the colors in the color scale}

\item{nmax}{Maximum number of different colors the palette should contain. If not provided, is calculated automatically from the data.}

\item{order}{Numeric vector listing the order in which the colors should be used. Default is 1:nmax.}

\item{...}{Additional arguments to pass to grDevices::colorRampPalette()}
}
\value{
Color palette function(n)
}
\description{
Return function(n) to interpolate a ttcolor palette
}
