% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttcolors.R
\docType{data}
\name{ttpalettes}
\alias{ttpalettes}
\title{A list of tt color palettes}
\format{
An object of class \code{list} of length 12.
}
\usage{
ttpalettes
}
\description{
Each palette is a list of pre-defined tt colors
}
\keyword{datasets}
