% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPR.R
\name{WPR}
\alias{WPR}
\title{William's \%R}
\usage{
WPR(HLC, n = 14, scale = FALSE)
}
\arguments{
\item{HLC}{Object that is coercible to xts or matrix and contains
High-Low-Close prices.  If only a univariate series is given, it will be
used.  See details.}

\item{n}{Number of periods to use.}

\item{scale}{Scale the result to be between 0 and -100.}
}
\value{
A object of the same class as \code{HLC} or a vector (if
\code{try.xts} fails) containing the William's \%R values.
}
\description{
William's \% R.
}
\details{
If an High-Low-Close series is provided, the indicator is calculated using
the high/low values.  If a vector is provided, the calculation only uses that
series.
}
\note{
The William's \%R calculation is similar to stochastics' fast \%K,
and the result of \code{WPR} is equal to \code{1-fastK}.

The value for William's \%R will be 0.5 whenever the highest high and
lowest low are the same over the last \code{n} periods.

William's \%R is usually scaled to be between 0 and -100, which is not what
\code{WPR} returns by default. Set \code{scale = TRUE} to return the result
with the usual scaling.
}
\examples{

 data(ttrc)
 hlc <- ttrc[,c("High","Low","Close")]
 stochOsc <- stoch(hlc)
 stochWPR <- WPR(hlc)

 # WPR is a transformation of stochastics' fastK
 all.equal(stochWPR, 1-stochOsc[,'fastK'])  # TRUE

 # WPR converted to the usual scaling between 0 and -100
 scaledWPR <- WPR(hlc, scale=TRUE)

 plot(tail(stochOsc[,"fastK"], 100), type="l",
     main="Fast \%K and Williams \%R", ylab="",
     ylim=range(cbind(stochOsc, stochWPR), na.rm=TRUE) )
 lines(tail(stochWPR, 100), col="blue")
 lines(tail(1-stochWPR, 100), col="red", lty="dashed")

}
\references{
The following site(s) were used to code/document this
indicator:\cr
\url{https://www.fmlabs.com/reference/WilliamsR.htm}\cr
\url{https://www.metastock.com/Customer/Resources/TAAZ/?p=126}\cr
\url{https://www.linnsoft.com/techind/williams-r-wpr}\cr
\url{https://school.stockcharts.com/doku.php?id=technical_indicators:williams_r}\cr
}
\seealso{
See \code{\link{stoch}}.
}
\author{
Joshua Ulrich
}
\keyword{ts}
