% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttcolors.R
\name{scale_fill_tt}
\alias{scale_fill_tt}
\title{Fill scale constructor for ttcolors}
\usage{
scale_fill_tt(
  palette = "main",
  aesthetics = "fill",
  discrete = TRUE,
  reverse = FALSE,
  nmax = NULL,
  order = NULL,
  ...
)
}
\arguments{
\item{palette}{Character name of palette. Run \code{names(ttr::ttpalettes)} for available options.}

\item{aesthetics}{Default to \code{color} for \code{\link[=scale_color_tt]{scale_color_tt()}} and \code{fill} for \code{\link[=scale_fill_tt]{scale_fill_tt()}}.}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not.}

\item{reverse}{If TRUE, reverse the order of the colors in the color scale.}

\item{nmax}{Maximum number of different colors the palette should contain. If not provided, is calculated automatically from the data.}

\item{order}{Numeric vector listing the order in which the colors should be used. Default is \code{1:nmax}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}} or
\code{\link[ggplot2:scale_gradient]{ggplot2::scale_color_gradientn()}}, used respectively when discrete is \code{TRUE} or \code{FALSE}.}
}
\description{
To be used in a ggplot, with option \code{discrete} set to \code{TRUE} for discrete values.
See \code{\link[=scale_color_tt]{scale_color_tt()}} for more.
}
