% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttthemes.R
\name{theme_tt}
\alias{theme_tt}
\title{ttecon theme}
\usage{
theme_tt(
  style = "slide",
  palette = "green",
  xaxis = TRUE,
  yaxis = FALSE,
  axis_color = NULL,
  xgrid = FALSE,
  ygrid = TRUE,
  legend = TRUE,
  axis_title = TRUE,
  axis_title_just = "rt"
)
}
\arguments{
\item{style}{Change the font family.
Defaults to "slide" for TeX Gyre Heros
Change to "paper" for Latin Modern Roman}

\item{palette}{Change the palette
Defaults to "green" for green-shaded lines
Change to "brown" for brown-shaded lines}

\item{xaxis}{x-axis line if \code{TRUE}
Defaults to \code{TRUE}
The line is colored "green" for a green palette and is colored "brown" for a brown palette.}

\item{yaxis}{y-axis line if \code{TRUE}
Defaults to \code{FALSE}
The line is colored "green" for a green palette and is colored "brown" for a brown palette.}

\item{axis_color}{Defines the color for axis titles, lines and text.
If \code{NULL} then use the palette color if style is "slide" and black otherwise.
Defaults to \code{NULL}}

\item{xgrid}{Turn on major x-grid-lines if \code{TRUE}
Defaults to \code{FALSE}
The line is colored "gradient green" for a green palette and is colored "gradient brown" for a brown palette.}

\item{ygrid}{Turn on major y-grid-lines if \code{TRUE}
Defaults to \code{TRUE}
The line is colored "gradient green" for a green palette and is colored "gradient brown" for a brown palette.}

\item{legend}{allow legend if \code{TRUE}
Defaults to \code{TRUE}}

\item{axis_title}{allow axis titles if \code{TRUE}
Defaults to \code{TRUE}}

\item{axis_title_just}{axis title font justification, one of \verb{[blmcrt]}
Defaults to \code{"rt"}}
}
\value{
ggplot theme
}
\description{
Theme for ttecon figures.
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(x = mpg * 0.43, y = wt * 0.4535924, colour = factor(cyl))) +
  geom_point(size = 2) +
  labs(
    title = "Car weight vs efficiency",
    subtitle = "Using sensible metrics",
    x = "Efficiency (km/l)",
    y = "Weight (1000 kg)",
    colour = "Cylinders",
    caption = "Brought to you by the letter 'T'"
  ) +
  theme_tt(style = "paper") +
  scale_color_tt()
}
