% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttthemes.R
\name{theme_tt}
\alias{theme_tt}
\title{ttecon theme}
\usage{
theme_tt(
  style = "slide",
  font = NULL,
  palette = "green",
  xaxis = TRUE,
  yaxis = FALSE,
  axis_text_color_style = "style",
  xgrid = FALSE,
  ygrid = TRUE,
  legend = FALSE,
  axis_title = TRUE,
  axis_title_just = "mm",
  xticks = FALSE,
  yticks = FALSE
)
}
\arguments{
\item{style}{Change the theme style.
Default to \code{"slide"} for TeX Gyre Heros (if \code{font} is \verb{NULL) and lighter axis title and text. Change to }"paper"\verb{for Latin Modern Roman (if}font\code{is}NULL) and darker axis title and text.}

\item{font}{Change the font family.
Default to \code{NULL} to use the defaults based on \code{style}.}

\item{palette}{Change the palette.
Default to \code{"green"} for green-shaded lines.
Change to \code{"brown"} for brown-shaded lines.}

\item{xaxis}{x-axis line if \code{TRUE}.
Default to \code{TRUE},
The line is colored "green" for a green palette and is colored "brown" for a brown palette.}

\item{yaxis}{y-axis line if \code{TRUE}.
Default to \code{FALSE}.
The line is colored "green" for a green palette and is colored "brown" for a brown palette.}

\item{axis_text_color_style}{define the color style for axis titles and text.
If \code{"style"} then use a saturated palette color, use the default colors otherwise.
Default to \code{"style"}.}

\item{xgrid}{Turn on major x-grid-lines if \code{TRUE}.
Default to \code{FALSE}.
The line is colored "gradient green" for a green palette and is colored "gradient brown" for a brown palette.}

\item{ygrid}{Turn on major y-grid-lines if \code{TRUE}.
Default to \code{TRUE}.
The line is colored "gradient green" for a green palette and is colored "gradient brown" for a brown palette.}

\item{legend}{Allow legend if \code{TRUE}.
Default to \code{FALSE}.}

\item{axis_title}{Allow axis titles if \code{TRUE}.
Default to \code{TRUE}.}

\item{axis_title_just}{Axis title font justification, one of \verb{[blmcrt]}
Default to \code{"mm"}.}

\item{xticks}{Add x-ticks if \code{TRUE}.
Default to \code{FALSE}.}

\item{yticks}{Add y-ticks if \code{TRUE}.
Default to \code{FALSE}.}
}
\value{
ggplot theme
}
\description{
Theme for ttecon figures.
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(x = mpg * 0.43, y = wt * 0.4535924, colour = factor(cyl))) +
  geom_point(size = 2) +
  labs(
    title = "Car weight vs efficiency",
    subtitle = "Using sensible metrics",
    x = "Efficiency (km/l)",
    y = "Weight (1000 kg)",
    colour = "Cylinders",
    caption = "Brought to you by the letter 'T'"
  ) +
  theme_tt(style = "paper") +
  scale_color_tt()
}
